:building_construction: Separeted control validators to validators page
all changes are functional
