+++ src/app/controllers/AppointmentController.js
@@ -1,4 +1,3 @@
-import * as Yup from 'yup';
 import { startOfHour, parseISO, isBefore, format, subHours } from 'date-fns';
 import pt from 'date-fns/locale/pt';
 import Appointment from '../models/Appointment';
@@ -11,13 +10,6 @@ import CancellationMail from '../jobs/CancellationMail';
 
 class AppointmentController {
   async store(req, res) {
-    const schema = Yup.object().shape({
-      provider_id: Yup.number().required(),
-      date: Yup.date().required(),
-    });
-    if (!(await schema.isValid(req.body))) {
-      return res.status(400).json({ error: 'Validation fails' });
-    }
     const { provider_id, date } = req.body;
     /*
      * check if provider_id is a provider
